feat(components/atom/badge): add custom types, refactor all the things. classes became unique
