+++ components/atom/badge/demo/index.js
@@ -1,28 +1,15 @@
-import AtomBadge, {
-  atomBadgeTypes,
-  atomBadgeDesigns,
-  atomBadgeSizes
-} from 'components/atom/badge/src'
-const {
-  Label,
-  H1,
-  H2,
-  H4,
-  Code,
-  Bold,
-  Paragraph,
-  Article,
-  UnorderedList,
-  ListItem,
-  DevIcon,
-  Grid,
-  Cell
-} = window.DOC
+import {H1, Paragraph} from '@s-ui/documentation-library'
+
+import ArticleType from './ArticleType'
+import ArticleCustomType from './ArticleCustomType'
+import ArticleDesign from './ArticleDesign'
+import ArticleSizes from './ArticleSizes'
+import ArticleContent from './ArticleContent'
+import ArticleDarkMode from './ArticleDarkMode'
 
 const BASE_CLASS_DEMO = `DemoAtomBadge`
 const CLASS_SECTION = `${BASE_CLASS_DEMO}-section`
 
-const icon = <DevIcon icon="DiGithubBadge" />
 const Demo = () => {
   return (
     <div className="sui-StudioPreview">
@@ -32,476 +19,17 @@ const Demo = () => {
           This is an informative UI element, it doesn't change on hover or
           click, so the user can't interact with it.
         </Paragraph>
-        <Article className={CLASS_SECTION}>
-          <H2>Type</H2>
-          <div>
-            <Paragraph>
-              These are the available <Code>types</Code> of bagdes, which are
-              solid by default.
-            </Paragraph>
-            <AtomBadge label="success" type={atomBadgeTypes.SUCCESS} />
-            <AtomBadge label="error" type={atomBadgeTypes.ERROR} />
-            <AtomBadge label="info" type={atomBadgeTypes.INFO} />
-            <AtomBadge label="alert" type={atomBadgeTypes.ALERT} />
-            <AtomBadge label="new" type={atomBadgeTypes.NEW} />
-            <AtomBadge label="neutral" type={atomBadgeTypes.NEUTRAL} />
-            <AtomBadge label="primary" type={atomBadgeTypes.PRIMARY} />
-            <AtomBadge label="secondary" type={atomBadgeTypes.SECONDARY} />
-          </div>
-          <Paragraph>___</Paragraph>
-          <div>
-            <Paragraph>
-              Use the Prop <Code>transparent</Code> to remove the background.
-            </Paragraph>
-            <AtomBadge
-              label="success"
-              type={atomBadgeTypes.SUCCESS}
-              transparent
-            />
-            <AtomBadge label="error" type={atomBadgeTypes.ERROR} transparent />
-            <AtomBadge label="info" type={atomBadgeTypes.INFO} transparent />
-            <AtomBadge label="alert" type={atomBadgeTypes.ALERT} transparent />
-            <AtomBadge label="new" type={atomBadgeTypes.NEW} transparent />
-            <AtomBadge
-              label="neutral"
-              type={atomBadgeTypes.NEUTRAL}
-              transparent
-            />
-            <AtomBadge
-              label="primary"
-              type={atomBadgeTypes.PRIMARY}
-              transparent
-            />
-            <AtomBadge
-              label="primary"
-              type={atomBadgeTypes.SECONDARY}
-              transparent
-            />
-          </div>
-        </Article>
+        <ArticleType className={CLASS_SECTION} />
         <br />
-        <Article className={CLASS_SECTION}>
-          <H2>Design</H2>
-          <div>
-            <Paragraph>
-              These are the available <Code>design</Code> types of bagdes, which
-              are solid by default.
-            </Paragraph>
-            <Grid cols={9} gutter={10} style={{width: 600}}>
-              <Cell>
-                <Label>Solid</Label>
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="success"
-                  type={atomBadgeTypes.SUCCESS}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="error"
-                  type={atomBadgeTypes.ERROR}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="info"
-                  type={atomBadgeTypes.INFO}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="alert"
-                  type={atomBadgeTypes.ALERT}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="new"
-                  type={atomBadgeTypes.NEW}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="neutral"
-                  type={atomBadgeTypes.NEUTRAL}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="primary"
-                  type={atomBadgeTypes.PRIMARY}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOLID}
-                  label="secondary"
-                  type={atomBadgeTypes.SECONDARY}
-                />
-              </Cell>
-              <Cell>
-                <Label>Soft</Label>
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="success"
-                  type={atomBadgeTypes.SUCCESS}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="error"
-                  type={atomBadgeTypes.ERROR}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="info"
-                  type={atomBadgeTypes.INFO}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="alert"
-                  type={atomBadgeTypes.ALERT}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="new"
-                  type={atomBadgeTypes.NEW}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="neutral"
-                  type={atomBadgeTypes.NEUTRAL}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="primary"
-                  type={atomBadgeTypes.PRIMARY}
-                />
-              </Cell>
-              <Cell>
-                <AtomBadge
-                  design={atomBadgeDesigns.SOFT}
-                  label="secondary"
-                  type={atomBadgeTypes.SECONDARY}
-                />
-              </Cell>
-            </Grid>
-          </div>
-        </Article>
+        <ArticleCustomType className={CLASS_SECTION} />
         <br />
-        <Article className={CLASS_SECTION}>
-          <H2>Sizes</H2>
-          <div>
-            <Paragraph>
-              There are three options for <Code>size</Code>
-            </Paragraph>
-            <Grid cols={2} gutter={10} style={{width: 100}}>
-              <Cell>
-                <Label>Small</Label>
-              </Cell>
-              <Cell>
-                <AtomBadge label="label" size={atomBadgeSizes.SMALL} />
-              </Cell>
-              <Cell>
-                <Label>Medium</Label>
-              </Cell>
-              <Cell>
-                <AtomBadge label="label" size={atomBadgeSizes.MEDIUM} />
-              </Cell>
-              <Cell>
-                <Label>Large</Label>
-              </Cell>
-              <Cell>
-                <AtomBadge label="label" size={atomBadgeSizes.LARGE} />
-              </Cell>
-            </Grid>
-          </div>
-        </Article>
+        <ArticleDesign className={CLASS_SECTION} />
         <br />
-        <Article className={CLASS_SECTION}>
-          <H2>Content</H2>
-          <H4>Text</H4>
-          <UnorderedList>
-            <ListItem>Numbers, text and icons are valid</ListItem>
-            <ListItem>The content will never split in two lines</ListItem>
-            <ListItem>There's no limit of characters</ListItem>
-          </UnorderedList>
-          <Grid cols={3} gutter={10} style={{width: 200}}>
-            <Cell>
-              <Label>Small</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge label="label" size={atomBadgeSizes.SMALL} />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                transparent
-                size={atomBadgeSizes.SMALL}
-              />
-            </Cell>
-            <Cell>
-              <Label>Medium</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge label="label" size={atomBadgeSizes.MEDIUM} />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                transparent
-                size={atomBadgeSizes.MEDIUM}
-              />
-            </Cell>
-            <Cell>
-              <Label>Large</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge label="label" size={atomBadgeSizes.LARGE} />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                transparent
-                size={atomBadgeSizes.LARGE}
-              />
-            </Cell>
-          </Grid>
-          <H4>Icons</H4>
-          <UnorderedList>
-            <ListItem>
-              <Bold>Icons are optional</Bold> and can be placed either on the
-              right or the left side, but never on both at the same time.
-            </ListItem>
-            <ListItem>
-              Small badges with background won't display icons
-            </ListItem>
-          </UnorderedList>
-          <Grid cols={5} gutter={10} style={{width: 500}}>
-            <Cell>
-              <Label>Small</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                size={atomBadgeSizes.SMALL}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                size={atomBadgeSizes.SMALL}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                transparent
-                size={atomBadgeSizes.SMALL}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                transparent
-                size={atomBadgeSizes.SMALL}
-              />
-            </Cell>
-            <Cell>
-              <Label>Medium</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                size={atomBadgeSizes.MEDIUM}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                size={atomBadgeSizes.MEDIUM}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                transparent
-                size={atomBadgeSizes.MEDIUM}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                transparent
-                size={atomBadgeSizes.MEDIUM}
-              />
-            </Cell>
-            <Cell>
-              <Label>Large</Label>
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                size={atomBadgeSizes.LARGE}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                size={atomBadgeSizes.LARGE}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconRight
-                transparent
-                size={atomBadgeSizes.LARGE}
-              />
-            </Cell>
-            <Cell>
-              <AtomBadge
-                label="label"
-                icon={icon}
-                iconLeft
-                transparent
-                size={atomBadgeSizes.LARGE}
-              />
-            </Cell>
-          </Grid>
-        </Article>
+        <ArticleSizes className={CLASS_SECTION} />
         <br />
-        <Article mode="dark">
-          <H2>Dark mode</H2>
-          <div>
-            <Paragraph>
-              There is no <Code>negative</Code> Prop for this component.
-            </Paragraph>
-            <Paragraph>
-              This is how the Badges look like over dark backgrounds.
-            </Paragraph>
-            <AtomBadge label="success" type={atomBadgeTypes.SUCCESS} />
-            <AtomBadge label="error" type={atomBadgeTypes.ERROR} />
-            <AtomBadge label="info" type={atomBadgeTypes.INFO} />
-            <AtomBadge label="alert" type={atomBadgeTypes.ALERT} />
-            <AtomBadge label="new" type={atomBadgeTypes.NEW} />
-            <AtomBadge label="neutral" type={atomBadgeTypes.NEUTRAL} />
-            <AtomBadge label="primary" type={atomBadgeTypes.PRIMARY} />
-            <AtomBadge label="secondary" type={atomBadgeTypes.SECONDARY} />
-          </div>
-          <Paragraph>___</Paragraph>
-          <div>
-            <AtomBadge
-              label="success"
-              type={atomBadgeTypes.SUCCESS}
-              transparent
-            />
-            <AtomBadge label="error" type={atomBadgeTypes.ERROR} transparent />
-            <AtomBadge label="info" type={atomBadgeTypes.INFO} transparent />
-            <AtomBadge label="alert" type={atomBadgeTypes.ALERT} transparent />
-            <AtomBadge label="new" type={atomBadgeTypes.NEW} transparent />
-            <AtomBadge
-              label="neutral"
-              type={atomBadgeTypes.NEUTRAL}
-              transparent
-            />
-            <AtomBadge
-              label="primary"
-              type={atomBadgeTypes.PRIMARY}
-              transparent
-            />
-            <AtomBadge
-              label="secondary"
-              type={atomBadgeTypes.SECONDARY}
-              transparent
-            />
-          </div>
-          <Paragraph>___</Paragraph>
-          <div>
-            <AtomBadge
-              label="success"
-              type={atomBadgeTypes.SUCCESS}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="error"
-              type={atomBadgeTypes.ERROR}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="info"
-              type={atomBadgeTypes.INFO}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="alert"
-              type={atomBadgeTypes.ALERT}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="new"
-              type={atomBadgeTypes.NEW}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="neutral"
-              type={atomBadgeTypes.NEUTRAL}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="primary"
-              type={atomBadgeTypes.PRIMARY}
-              design={atomBadgeDesigns.SOFT}
-            />
-            <AtomBadge
-              label="secondary"
-              type={atomBadgeTypes.SECONDARY}
-              design={atomBadgeDesigns.SOFT}
-            />
-          </div>
-        </Article>
+        <ArticleContent className={CLASS_SECTION} />
+        <br />
+        <ArticleDarkMode className={CLASS_SECTION} />
       </div>
     </div>
   )
