+++ components/atom/badge/demo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "atom-badge",
+  "name": "@s-ui/react-atom-badge-demo",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
